+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name =  tricc
 version = 0.0.3
-author = Rafael Klünder & DHU
+author = DHU SwissTPH
 author_email = httu.admin@swisstph.ch
 description = Python library that converts XLS to Fhir SDC ressource.
 long_description = file: README.md
