+++ tricc/strategies/output/xlsform_cht.py
@@ -7,7 +7,7 @@ import pandas as pd
 
 from tricc.models.lang import SingletonLangClass
 from tricc.serializers.xls_form import SURVEY_MAP
-from tricc.strategies.xlsform_cdss import XLSFormCDSSStrategy
+from tricc.strategies.output.xlsform_cdss import XLSFormCDSSStrategy
 
 langs = SingletonLangClass()
